add: base app home controller
